+++ lib/src/io/commands.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/args.dart';
@@ -5,7 +6,8 @@ import 'package:args/command_runner.dart';
 import 'package:args/src/arg_parser.dart';
 import 'package:dcli/dcli.dart';
 import 'package:flutter_translation_sheet/flutter_translation_sheet.dart';
-
+import 'package:yaml/yaml.dart';
+import 'package:path/path.dart' as p;
 import '../runner.dart';
 
 /// Command logic for `fts extract`
@@ -73,6 +75,7 @@ class FetchCommand extends Command<int> {
 
   @override
   Future<int> run() async {
+    readPubSpec();
     setConfig(argResults!);
     await exec();
     return 0;
@@ -101,7 +104,7 @@ class UpgradeCommand extends Command<int> {
 class RunCommand extends Command<int> {
   @override
   final String description =
-      'Runs the strings parsing, starts the sync process with GoogleSheet, fetches the translations, and generates the files.';
+      '(default) Runs the strings parsing, starts the sync process with GoogleSheet, fetches the translations, and generates the files.';
 
   @override
   final String name = 'run';
@@ -190,14 +193,76 @@ void setConfig(ArgResults res) {
   if (res.wasParsed('config')) {
     startConfig(res['config']);
   } else {
-    startConfig('trconfig.yaml');
+    readPubSpec();
+    var ftsKey = pubSpecMap['fts'];
+    if (ftsKey != null) {
+      if (ftsKey is String) {
+        /// path.
+        startConfig(ftsKey);
+      } else if (ftsKey is YamlMap) {
+        /// map.
+        loadEnv(parsedDoc: ftsKey);
+      } else {
+        error("Invalid `fts` key in pubspec.yaml");
+        exit(1);
+      }
+    } else {
+      startConfig('trconfig.yaml');
+    }
   }
+}
+
+String pubSpecStr = '';
+Map pubSpecMap = {};
+
+/// Adds the json files directory. to pubspec.yaml assets.
+void addAssetsToPubSpec() {
+  if (pubSpecStr.isEmpty) {
+    readPubSpec();
+  }
+  var addAsset = p.dirname(config.outputJsonTemplate) + '/';
+  var out = pubSpecStr;
+  var assets = pubSpecMap['flutter']?['assets'];
+  var replacer = '';
+  if (assets == null) {
+    replacer = kNoAssetsReplace.replaceAll('##replace', addAsset);
+    out = out.replaceAll(kNoAssetsKey, replacer);
+  } else {
+    // check if the key is already added
+    var hasAsset = false;
+    if (assets is List) {
+      for (String asset in assets) {
+        if (asset.contains(addAsset)) {
+          hasAsset = true;
+        }
+      }
+    }
+    if (!hasAsset) {
+      replacer = kHasAssetsReplace.replaceAll('##replace', addAsset);
+      out = out.replaceAll(kHasAssetsKey, replacer);
+    }
+  }
+  if (out != pubSpecStr) {
+    saveString('pubspec.yaml', out);
+  }
+}
+
+/// Reads the pubspec.yaml file and stores it in [pubSpecStr] and [pubSpecMap]
+void readPubSpec() {
+  final file = File('pubspec.yaml');
+  if (!file.existsSync()) {
+    error("Can't locate pubspec.yaml, run fts from your project root.");
+    exit(3);
+  }
+  pubSpecStr = openString(file.path);
+  pubSpecMap = loadYaml(pubSpecStr) as Map;
 }
 
 /// Initializes the supplied configuration from [path]
 void startConfig(String path) {
   if (path.isEmpty) {
-    error('Pass the trconfig.yaml path to -c');
+    error(
+        'Pass the trconfig.yaml path to --config, or add `fts` to pubspec.yaml');
     exit(1);
   }
   var f = File(path);
@@ -207,7 +272,7 @@ void startConfig(String path) {
     /// ask to create from template.
     var useCreateTemplate = confirm(
         yellow(
-            'Do you wanna create the template trconfig.yaml in the current directory?'),
+            'Do you wanna create the template (trconfig.yaml) in the current directory?'),
         defaultValue: true);
     if (!useCreateTemplate) {
       var m1 = grey('${CliConfig.cliName} run', background: AnsiColor.black);
@@ -222,6 +287,8 @@ void startConfig(String path) {
       createSampleContent();
     }
   } else {
-    loadEnv(path);
+    loadEnv(
+      path: path,
+    );
   }
 }